# services/gateway/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Gateway Service API",
    "description": "Gateway service that proxies requests to downstream services with JWT authentication",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "API Support",
      "url": "http://localhost:8080"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
    "/info": {
      "get": {
        "summary": "Health check",
        "description": "Returns the health status of the gateway and all downstream services",
        "operationId": "getInfo",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string"
                },
                "services": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "description": "Creates a new user account and returns JWT token in cookie",
        "operationId": "register",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "minLength": 3
                },
                "password": {
                  "type": "string",
                  "minLength": 8
                },
                "phone": {
                  "type": "string",
                  "minLength": 6
                },
                "email": {
                  "type": "string",
                  "format": "email"
                }
              },
              "required": [
                "username",
                "password",
                "phone"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User registered successfully",
            "headers": {
              "Set-Cookie": {
                "type": "string",
                "description": "JWT token in httpOnly cookie"
              }
            },
            "schema": {
              "type": "object",
              "properties": {
                "access_token": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Phone number or username already taken",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login user",
        "description": "Authenticates user and returns JWT token in cookie",
        "operationId": "login",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              },
              "required": [
                "username",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful",
            "headers": {
              "Set-Cookie": {
                "type": "string",
                "description": "JWT token in httpOnly cookie"
              }
            },
            "schema": {
              "type": "object",
              "properties": {
                "access_token": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/conversations": {
      "post": {
        "summary": "Create a new conversation",
        "operationId": "createConversation",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Conversation created successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "summary": "Get conversations for user",
        "operationId": "getUserConversations",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of conversations"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/conversations/{id}": {
      "get": {
        "summary": "Get conversation details",
        "operationId": "getConversation",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation details"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "summary": "Update conversation",
        "operationId": "updateConversation",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation updated"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "summary": "Delete conversation",
        "operationId": "deleteConversation",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Conversation deleted"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/messages": {
      "post": {
        "summary": "Create a new message",
        "operationId": "createMessage",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Message created"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/messages/{id}": {
      "get": {
        "summary": "Get message details",
        "operationId": "getMessage",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Message details"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "summary": "Update message",
        "operationId": "updateMessage",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Message updated"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "summary": "Delete message",
        "operationId": "deleteMessage",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Message deleted"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "CookieAuth": {
      "type": "apiKey",
      "name": "access_token",
      "in": "cookie"
    }
  }
}